Add Windows-specific compiler arguments to Fortran object in meson.build
This commit refines the handling of Windows-specific compiler arguments in the meson.build file by ensuring that the '-D__MINGW_USE_VC_SETJMP=1' argument is included in the static library definition for the Fortran object. This change enhances compatibility with MinGW's setjmp/longjmp functionality, further improving the build process for Windows environments.